split PsModule out of models
