+++ apps/exploits/serializers.py
@@ -1,7 +1,8 @@
 from rest_framework import serializers
 
 from apps.exploits.models import EPSS, Exploit
-from osidb.models import Affect, Flaw, PsModule, Tracker
+from osidb.dmodels import PsModule
+from osidb.models import Affect, Flaw, Tracker
 
 
 class ExploitsSimpleSerializer(serializers.ModelSerializer):
