+++ osidb/dmodels/__init__.py
@@ -1,3 +1,4 @@
+from .ps_module import PsModule
 from .ps_product import PsProduct
 
-__all__ = ("PsProduct",)
+__all__ = ("PsModule", "PsProduct")
