+++ osidb/models.py
@@ -46,8 +46,7 @@ from apps.trackers.models import JiraBugIssuetype
 from apps.workflows.workflow import WorkflowFramework, WorkflowModel
 from collectors.bzimport.constants import BZ_API_KEY, FLAW_PLACEHOLDER_KEYWORD
 
-from .dmodels import PsProduct
-from .helpers import ps_update_stream_natural_keys
+from .dmodels import PsModule
 from .mixins import (
     ACLMixin,
     ACLMixinManager,
@@ -3229,97 +3228,6 @@ class PackageVer(models.Model):
         verbose_name = "Version"
 
     version = models.CharField(max_length=1024)
-
-
-class PsModule(NullStrFieldsMixin, ValidateMixin):
-
-    # internal primary key
-    uuid = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-
-    # name of the module
-    name = models.CharField(max_length=100, unique=True)
-
-    public_description = models.TextField()
-    cpe = fields.ArrayField(models.TextField(), default=list, blank=True)
-
-    # Flags
-    private_trackers_allowed = models.BooleanField(default=False)
-    autofile_trackers = models.BooleanField(default=False)
-    special_handling_features = fields.ArrayField(
-        models.TextField(), default=list, blank=True
-    )
-
-    # BTS
-    bts_name = models.CharField(max_length=50)
-    bts_key = models.CharField(max_length=100)
-    bts_groups = models.JSONField(default=dict)
-
-    # Lifecycle
-    supported_from_dt = models.DateTimeField(null=True, blank=True)
-    supported_until_dt = models.DateTimeField(null=True, blank=True)
-
-    # CC Lists
-    default_cc = fields.ArrayField(
-        models.CharField(max_length=50), default=list, blank=True
-    )
-    component_cc = models.JSONField(default=dict, blank=True)
-    private_tracker_cc = fields.ArrayField(
-        models.CharField(max_length=50), default=list, blank=True
-    )
-
-    # Component overrides
-    default_component = models.CharField(max_length=100, blank=True)
-    component_overrides = models.JSONField(default=dict, blank=True)
-
-    # Update Streams
-    # implicit:
-    # ps_update_streams
-    # active_ps_update_streams
-    # default_ps_update_streams
-    # aus_ps_update_streams
-    # unacked_ps_update_stream
-
-    ps_product = models.ForeignKey(
-        PsProduct, on_delete=models.CASCADE, related_name="ps_modules"
-    )
-
-    @property
-    def is_rhscl(self) -> bool:
-        """
-        check and return whether the PS module is RHSCL one
-
-        Red Hat Software Collections represent an extra layer in
-        the component hierarchy and may require special handling
-        """
-        return self.bts_key == RHSCL_BTS_KEY
-
-    @property
-    def y_streams(self):
-        """Current Y-stream(s) - it can be more of them"""
-
-        return list(self.active_ps_update_streams.exclude(name__endswith="z"))
-
-    @property
-    def z_stream(self):
-        """Current Z-stream"""
-        z_streams = self.active_ps_update_streams.filter(name__endswith="z")
-        return (
-            max(list(z_streams), key=ps_update_stream_natural_keys)
-            if z_streams
-            else None
-        )
-
-    def subcomponent(self, component) -> Union[str, None]:
-        """
-        return the subcomponent for the given component or None if not present
-        """
-        if (
-            self.component_overrides
-            and component in self.component_overrides
-            and isinstance(self.component_overrides[component], dict)
-            and self.component_overrides[component].get("sub_component")
-        ):
-            return self.component_overrides[component]["sub_component"]
 
 
 class PsUpdateStream(NullStrFieldsMixin, ValidateMixin):
